fabric: upstream runif to pkg
