+++ src/lightning/fabric/utilities/testing/__init__.py
@@ -1,0 +1,3 @@
+from lightning.fabric.utilities.testing._runif import _RunIf
+
+__all__ = ["_RunIf"]
